+++ src/components/expand-viewer/expand-viewer.styles.ts
